requirements.txt
flask
pandas
numpy
scikit-learn
matplotlib
seaborn
wordcloud
python-dotenv
